[mod] 패시브 아이템 스테이터스 적용

+++ SurviveGame/Assets/Scripts/Class/Model/Character.cs
@@ -42,7 +42,7 @@ public class Character
     }
 
 
-    public void UpdataStatus(CharacterStatusType _type, int _amount)
+    public void UpdataStatus(CharacterStatusType _type, float _amount)
     {
         statusArr[(int)_type].UpdataStatus(_amount);
     }
